feat: Add Swiper npm dependency for image carousel
Add the Swiper npm dependency to the ecoEssence front-end project. This dependency is required for implementing an image carousel feature on certain pages of the website. The Swiper library provides a user-friendly and interactive way for users to browse through a collection of images. This addition enhances the visual appeal and functionality of the website.