feat: Add image carousel component to the homepage
The commit adds a new component called Carousel to the ecoEssence front-end project. This component is responsible for rendering an image carousel on the homepage of the website. It uses the Swiper library for the carousel functionality and includes CSS styles for the carousel layout. This addition enhances the visual appeal and user experience of the homepage.