feat: Add blog post grid to the Blog page(demo with dummy data will be replaced later)
The commit adds a blog post grid to the Blog page in the ecoEssence front-end project. This grid displays a collection of blog posts, including their titles, dates, descriptions, and images. It enhances the visual appeal and user experience of the Blog page by providing an organized and engaging way for users to browse and read blog posts.